pyproject.toml
[project]
name = "ideographco"
version = "0.1.0"
description = "Sistema de regresión multisalida para cuantificación de ideologías políticas en noticias colombianas"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.30",
    "pytorch-lightning>=2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "plotly>=5.0",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "newspaper4k>=0.9",
    "feedparser>=6.0",
    "torchmetrics>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "ipykernel>=6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["src*"]
